+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Handle variables with duplicate dimensions ([#299](https://github.com/nasa/stitchee/pull/299))([**@ank1m**](https://github.com/ank1m))
+- Handle non-concatenating variables of the same name for potential conflict ([#299](https://github.com/nasa/stitchee/pull/299))([**@ank1m**](https://github.com/ank1m))
+
 ## [1.7.0] - 2025-07-07
 
 ### Changed